🦄refactor(mac/tray/lyric): 解耦 macOS 状态栏歌词至独立模块
- 将 macOS 状态栏歌词功能从  模块完全迁移至新增的  独立模块，显著提升模块化和可维护性。
-  现已专用于处理 Windows/Linux 平台的任务栏歌词逻辑，消除了平台间的耦合。
-  更新为根据操作系统平台动态加载  或 。
- 调整了  对  事件的响应逻辑，并移除了冗余的  监听器，确保托盘标题管理与歌词功能的独立控制。
- 歌词和播放状态数据（如  等）的底层 IPC 事件保持共享，无需修改渲染进程的数据发送逻辑。
