+++ electron/main/ipc/index.ts
@@ -1,6 +1,8 @@
+import { isMac } from "../utils/config";
 import initCacheIpc from "./ipc-cache";
 import initFileIpc from "./ipc-file";
 import initLyricIpc from "./ipc-lyric";
+import { initMacStatusBarIpc } from "./ipc-mac-statusbar";
 import initMediaIpc from "./ipc-media";
 import initMpvIpc from "./ipc-mpv";
 import initProtocolIpc from "./ipc-protocol";
@@ -35,7 +37,11 @@ const initIpc = (): void => {
   initMediaIpc();
   initMpvIpc();
   initRendererLogIpc();
-  initTaskbarIpc();
+  if (isMac) {
+    initMacStatusBarIpc();
+  } else {
+    initTaskbarIpc();
+  }
 };
 
 export default initIpc;
